landslide-edge-simulation/simulator/requirements.txt
paho-mqtt>=2.0.0

